requirements/production.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.8

# App server
gunicorn

# Templates
django_compressor

# Security
bleach
#cef

# Logging
#raven

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Admin
django-debug-toolbar

# Feature testing
#django-waffle

# environment
django-environ

django-bootstrap3==6.2.2

django-braces==1.8.1
